Make it safer parsing time and satellites
In my calls to gpsd using this library, I am getting no
satellites and the time is in UTC, which includes Z.  this
patch makes it safer to parse the time format so it doesn't fail.

Also, in my data coming from gpsd it isn't reporting satellites,
but correctly reports gps lat/lon/az.  this patch makes sure
there are actual satellites in the data before trying to walk them.